+++ geoextent/lib/content_providers/Dataverse.py
@@ -16,6 +16,7 @@ Supported Dataverse installations:
 - Recherche Data Gouv (recherche.data.gouv.fr)
 - ioerDATA (data.fdz.ioer.de)
 - heiDATA (heidata.uni-heidelberg.de)
+- Edmond (edmond.mpg.de)
 - Other Dataverse installations following the standard API
 
 Supported identifier formats:
@@ -63,6 +64,7 @@ class Dataverse(DoiProvider):
             "recherche.data.gouv.fr",
             "data.fdz.ioer.de",
             "heidata.uni-heidelberg.de",
+            "edmond.mpg.de",
         ]
 
         # URL patterns for validation
@@ -185,6 +187,7 @@ class Dataverse(DoiProvider):
             r"10\.5064/",  # UNC Dataverse
             r"10\.71830/",  # ioerDATA (Leibniz Institute of Ecological Urban and Regional Development)
             r"10\.11588/DATA/",  # heiDATA (Heidelberg University)
+            r"10\.17617/",  # Edmond (Max Planck Society)
         ]
 
         return any(